Change ProductBox into wrapper for product display components
