+++ src/components/common/ProductBox/GridProduct/GridProduct.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/settings.scss";
+@import "../../../../styles/settings.scss";
 
 .root {
   background-color: $pb-bcg;
